Console refactor
