+++ app/controllers/pages_controller.rb
@@ -24,23 +24,6 @@ class PagesController < ApplicationController
     @locale   = set_locale
   end
 
-  def console
-    unless current_user.role == "admin" 
-      render_404
-    end
-    #Link to chartkick tutorial: https://gorails.com/episodes/charts-with-chartkick-and-groupdate
-  end
-
-  private
-
-  def render_404
-    respond_to do |format|
-      format.html { render :file => "#{Rails.root}/public/404.html", :layout => false, :status => :not_found }
-      format.xml  { head :not_found }
-      format.any  { head :not_found }
-    end
-  end
-
   def set_base_url
     url = request.url
     uri = URI.parse(url)
