+++ app/helpers/console_helper.rb
@@ -1,0 +1,2 @@
+module ConsoleHelper
+end
