+++ app/views/console/index.html.erb
