+++ app/views/shared/_navbar.html.erb
@@ -39,7 +39,7 @@
             <p><%= link_to t("pilot_prefs.menu"), edit_pilot_pref_path(current_user) %></p>
             <p><%= link_to t("shared.sign_out"), destroy_user_session_path, data: {turbo_method: :delete} %></p>
             <% if current_user.role == "admin" %>
-              <p><%= link_to t("shared.console"), pages_console_path %></p>
+              <p><%= link_to t("shared.console"), console_index_path %></p>
             <% end %>
           <% else %>
             <p><%= link_to t("shared.sign_in"), new_user_session_path(redirect_to: request.path) %></p>
@@ -57,7 +57,7 @@
           <p><%= link_to t("pilot_prefs.menu"), edit_pilot_pref_path(current_user) %></p>
           <p><%= link_to t("shared.sign_out"), destroy_user_session_path, data: {turbo_method: :delete} %></p>
           <% if current_user.role == "admin" %>
-            <p><%= link_to t("shared.console"), pages_console_path %></p>
+            <p><%= link_to t("shared.console"), console_index_path %></p>
           <% end %>
         <% else %>
           <p><%= link_to t("shared.sign_in"), new_user_session_path(redirect_to: request.path) %></p>
